Refactor to prepare for delivery
