+++ src/OPTest/classes/__init__.py
@@ -1,0 +1,2 @@
+from .OPTestAPIv2 import OPTestAPIv2
+from .OPTestGRCObject import OPTestGRCObject
