+++ src/OPTest/test/test_wf_SOXIssue.py
@@ -1,8 +1,15 @@
 
+import sys
 import os
+
+BASE_DIR = os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
+sys.path.append(BASE_DIR)
+
 from OPTest import OPTestAPIv2
 from OPTest import OPTestGRCObject
+
 import unittest
+
 
 def setUpModule():
     op_url = os.environ['OP_URL']
